requirements.txt
streamlit>=1.32.0
langgraph>=0.0.60
langchain-core>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
cachetools>=5.3.0
langdetect>=1.0.9
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
prometheus-client>=0.20.0
